Slater milestone2

+++ MonoGame.Framework/Graphics/PackedVector/Alpha8.cs
@@ -44,7 +44,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
         /// <returns>The packed vector in Vector3 format</returns>
         public float ToAlpha()
         {
-            return (float) (packedValue / 255.0f);
+            return (packedValue / 255.0f);
         }
 
         /// <summary>
@@ -66,7 +66,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
                 0.0f,
                 0.0f,
                 0.0f,
-                (float) (packedValue / 255.0f)
+(packedValue / 255.0f)
             );
         }
 
@@ -77,7 +77,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
         /// <returns>True if the object is equal to the packed vector.</returns>
         public override bool Equals(object obj)
         {
-            return (obj is Alpha8) && Equals((Alpha8) obj);
+            return (obj is Alpha8) && Equals((Alpha8)obj);
         }
 
         /// <summary>
@@ -120,7 +120,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
 
         private static byte Pack(float alpha)
         {
-            return (byte) Math.Round(
+            return (byte)Math.Round(
                 MathHelper.Clamp(alpha, 0, 1) * 255.0f
             );
         }
