+++ MonoGame.Framework/Graphics/PackedVector/HalfTypeHelper.cs
@@ -97,7 +97,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
                         mantissa = mantissa << 1;
                     }
                     mantissa &= 0xfffffbff;
-                    rst = ((uint)((((uint)value & 0x8000) << 16) | ((exp + 127) << 23))) | (mantissa << 13);
+                    rst = ((((uint)value & 0x8000) << 16) | ((exp + 127) << 23)) | (mantissa << 13);
                 }
                 else
                 {
@@ -106,7 +106,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
             }
             else
             {
-                rst = (uint)(((((uint)value & 0x8000) << 16) | ((((((uint)value >> 10) & 0x1f) - 15) + 127) << 23)) | (mantissa << 13));
+                rst = (((((uint)value & 0x8000) << 16) | ((((((uint)value >> 10) & 0x1f) - 15) + 127) << 23)) | (mantissa << 13));
             }
 
             var uif = new uif();
